Add functionality to disable Watchdog killing System Server
Introduced a way to control the behavior of Android WatchDog for
any system process if its stuck.

This is needed for live debugging of user space issues.
Otherwise Android Watch Dog will kill the process that is
stuck for Monitor time.

Before watchdog kill system_server, we will get the tombstones of
system_server.

-Added logic to obtain the WDOG behavior based on System Properties.

-Now sends SIGSTOP once the System property is set.

(cherry picked from commit 728d6fd86582bf58bc1febaecf19782881ea0ad5)

+++ services/java/com/android/server/Watchdog.java
@@ -395,7 +395,6 @@ public class Watchdog extends Thread {
         while (true) {
             mCompleted = false;
             mHandler.sendEmptyMessage(MONITOR);
-
             synchronized (this) {
                 long timeout = TIME_TO_WAIT;
 
@@ -473,10 +472,19 @@ public class Watchdog extends Thread {
             } catch (InterruptedException ignored) {}
 
             // Only kill the process if the debugger is not attached.
-            if (!Debug.isDebuggerConnected()) {
-                Slog.w(TAG, "*** WATCHDOG KILLING SYSTEM PROCESS: " + name);
-                Process.killProcess(Process.myPid());
-                System.exit(10);
+            if(!Debug.isDebuggerConnected()) {
+               if(SystemProperties.getInt("sys.watchdog.disabled", 0) == 0) {
+
+                  Slog.w(TAG, "*** WATCHDOG KILLING SYSTEM PROCESS: " + name);
+                  Process.killProcess(Process.myPid());
+                  System.exit(10);
+                }
+                else {
+                  Slog.w(TAG, "*** WATCHDOG NOT KILLING SYSTEM PROCESS: " + name);
+                  // Send the SIGSTOP to the System Process for DEBUG purposes
+                  Process.sendSignal(Process.myPid(), 19);
+                  break;
+                }
             } else {
                 Slog.w(TAG, "Debugger connected: Watchdog is *not* killing the system process");
             }
